[solidjs-kit] Header not showing on refresh.
