+++ solidjs-tailwind/src/App.jsx
@@ -5,11 +5,11 @@ import ROUTES from './routes';
 
 function App() {
   return (
-   <>
+    <>
       <Routes>
         <Route component={SigninPage} path={ROUTES.SIGNIN} />
         <Route component={RedirectPage} path={ROUTES.REDIRECT} />
-        <Route component={AuthGuard} path="/" >
+        <Route component={AuthGuard} path="/">
           <Route component={Home} path={ROUTES.HOME} />
           <Route component={Profile} path={ROUTES.PROFILE} />
           <Route component={OrgProfile} path={ROUTES.ORGPROFILE} />
